requirements.txt
# requirements.txt - GitHub Actions compatible
google-api-python-client==2.145.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.8.1
gspread==6.1.2
oauth2client==4.1.3
moviepy==1.0.3
python-dotenv==1.0.0
numpy==1.26.4
Pillow==10.3.0
requests==2.32.3
tenacity==8.4.2
python-dateutil==2.9.0.post0
pytz==2024.1